+++ CHANGELOG.md
@@ -112,6 +112,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   example. ([#266])
 - Adc's `SampleTime` type has been reworked and is now a consistent wrapper around
   the underlying types for `stm32f3`'s `SMP9_A` and `SMP18_A` type. ([#266])
+- Rename `CkMode` to `ClockMode` ([#266])
 
 ## [v0.7.0] - 2021-06-18